Implement else if expressions
