+++ evaluator/evaluator_test.go
@@ -100,6 +100,9 @@ func TestIfElseExpressions(t *testing.T) {
 		{"if (1 > 2) { 10 }", nil},
 		{"if (1 > 2) { 10 } else { 20 }", 20},
 		{"if (1 < 2) { 10 } else { 20 }", 10},
+		{"if (1 < 2) { 10 } else if (1 == 1) { 20 } else { 30 }", 10},
+		{"if (1 > 2) { 10 } else if (1 == 1) { 20 } else { 30 }", 20},
+		{"if (1 > 2) { 10 } else if (1 == 2) { 20 } else { 30 }", 30},
 	}
 
 	for _, tt := range tests {
